MSE800-PSE/WEEK10/BACKEND_KAIWHAKIRITE/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-cors==4.0.0
flask-jwt-extended==4.5.2
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
SQLAlchemy==2.0.20
